refactored tests with TestUtils
